:bug: #2973 【小程序】修改获取url link接口请求类中的expireTime属性类型为Long

+++ weixin-java-miniapp/src/main/java/cn/binarywang/wx/miniapp/bean/urllink/GenerateUrlLinkRequest.java
@@ -74,7 +74,7 @@ public class GenerateUrlLinkRequest implements Serializable {
    * </pre>
    */
   @SerializedName("expire_time")
-  private Integer expireTime;
+  private Long expireTime;
 
   /**
    * 到期失效的URL Link的失效间隔天数。生成的到期失效URL Link在该间隔时间到达前有效。最长间隔天数为365天。expire_type 为 1 必填
